# apps/docs/openapi/writing-app-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Writing App API",
    "description": "개발 문서",
    "version": "0.0.1"
  },
  "paths": {
...
    "/courses/{courseId}": {
      "get": {
        "operationId": "getCoursesByCourseId",
        "responses": {
          "200": {
            "description": "코스 상세입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1
                    },
                    "description": {
                      "type": "string",
                      "minLength": 1
                    },
                    "thumbnail": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "firstLessonId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "chapters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "label": {
                            "type": "string",
                            "minLength": 1
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1
                          },
                          "lessons": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "lessonId": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "title": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "description": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "order": {
                                  "type": "integer",
                                  "exclusiveMinimum": 0,
                                  "maximum": 9007199254740991
                                }
                              },
                              "required": [
                                "id",
                                "lessonId",
                                "title",
                                "description",
                                "order"
                              ]
                            }
                          }
                        },
                        "required": ["id", "label", "title", "lessons"]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "description",
                    "thumbnail",
                    "lessonCount",
                    "chapters"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "코스를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "course-not-found"
                    },
                    "message": {
                      "type": "string",
                      "const": "코스를 찾을 수 없습니다."
                    },
                    "courseId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message", "courseId"]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "courseId",
            "required": true
          }
        ]
      }
    },
...
  },
...
}